fix(gap-sweep): address PR #575 Copilot bot inline review (2 findings)
Two correctness bugs in the list-display labels of EDFE7ViewModel:

1. LoadLynList() formatted the UnitId using rom.u8(addr) but the
   Lyn record's D0 field is a DWord (WF EDFE7Form.N3_Init uses
   Program.ROM.u32(addr)). Reading u8 would drop the upper 3
   bytes and diverge from the u32==0 terminator predicate.
   Changed to rom.u32(addr).

2. LoadEpithetList() had the same issue - FE7's epithet record
   stores UnitId as a DWord at +0 (WF EDFE7Form.N1_Init uses
   Program.ROM.u32(addr)). Changed to rom.u32(addr).

Tests pass: 37 EDFE7 parity tests still green (the synthetic
ROMs use low UnitId values where u8 / u32 happen to match,
but the change matches the WF source-of-truth).

User prompt: Drive GitHub issue #403 (gap-sweep) on
laqieer/FEBuilderGBA from OPEN -> MERGED.

+++ FEBuilderGBA.Avalonia/ViewModels/EDFE7ViewModel.cs
@@ -186,7 +186,11 @@ namespace FEBuilderGBA.Avalonia.ViewModels
                 LYN_BLOCK_SIZE,
                 (addr) => rom.u32(addr) != 0x00,
                 (addr) => {
-                    uint uid = rom.u8(addr);
+                    // FE7 Lyn record's UnitId is a DWord at +0 (WF N3_Init
+                    // uses Program.ROM.u32(addr)). Reading u8 here would
+                    // drop the upper 3 bytes and diverge from the u32==0
+                    // terminator predicate above.
+                    uint uid = rom.u32(addr);
                     return $"{U.ToHexString(uid)} {GetUnitNameForUid(uid)}";
                 });
         }
@@ -302,7 +306,11 @@ namespace FEBuilderGBA.Avalonia.ViewModels
                 EPITHET_BLOCK_SIZE,
                 (addr) => rom.u32(addr) != 0x00,
                 (addr) => {
-                    uint uid = rom.u8(addr);
+                    // FE7 epithet record's UnitId is a DWord at +0 (WF
+                    // N1_Init uses Program.ROM.u32(addr)). Reading u8
+                    // would drop the upper 3 bytes and diverge from the
+                    // u32==0 terminator predicate above.
+                    uint uid = rom.u32(addr);
                     uint textId = rom.u32(addr + 4);
                     string textPreview = textId != 0 ? NameResolver.GetTextById(textId) : "";
                     return $"{U.ToHexString(uid)} {GetUnitNameForUid(uid)} {textPreview}".Trim();
